backend/requirements.txt
# Python dependencies untuk Sistem UKS Sekolah Backend
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0